(PUP-8066) Updated a set of acceptance tests for fips compliance
The tests have been adjusted to detect fips mode using a
fips fact so they can execute on both fips and non-fips platforms.

Updated versions of acceptance tests to run in FIPS mode:
 - tests/puppet_apply_a_file_should_create_a_file_and_report_the_md5.rb
 - tests/ticket_1334_clientbucket_corrupted.rb
 - tests/ticket_6541_invalid_filebucket_files.rb
 - tests/parser_functions/calling_all_functions.rb
 - tests/resource/file/content_attribute.rb

Also contains following change to setup/pre-suite/010_Install.rb
  - To only adjust digest_algorithm on fips agents
which is needed only for testing purposes till we have a FIPS
enabled vmpooler image.

+++ acceptance/tests/modules/build/build_agent.rb
@@ -4,6 +4,12 @@ tag 'audit:medium',
     'audit:acceptance'
 
 agents.each do |agent|
+
+  if (on(agent, facter("find in_fips_mode")).stdout =~ /true/)
+    puts "Module build, loading and installing not supported on fips enabled platforms"
+    next
+  end
+
   teardown do
     on agent, 'rm -rf bar'
   end
